Not Happy with Current Course?
UNABLE TO COPE UP with YOUR current course?

Social WeScholarship · Add Credential
Answered Wed

UNABLE TO COPE UP WITH YOUR CURRENT COURSE?

Some of you might find that you are not as happy with the course you chose as you expected to be. If this is you, then we recommend you consider We Scholarship!

On the off chance that you can't change to a course that you might want to study or find that the college you have picked just isn't for you, at that point pulling back from college can be a reasonable arrangement. It is significant that you are content with your choice so don't rush.

If you withdraw from the university then depending on how much of your current program you have completed, you may be eligible to receive some kind of award, so discuss this with your course coordinator.

Ensure that what will happen to any student loan that you have received as well as any fees for tuition or accommodation that you may have already paid or be liable for.

In the event that you do need to pull back from college don't let this put you off from continuing your studies. At the point when you discover the course, you might want to consider applying as a new student, however your experience will better place you to settle on the right decision.

This unhappiness forces the International students to think that – AM I UNABLE TO COPE UP WITH YOUR CURRENT COURSE? And we are sure that the answer is “YES”

Here, We Scholarship steps in to help you come over this tensed situation. At We Scholarship, we have confidence in making a superior regular day to day existence for International Students and diminish the budgetary weight on them, so they can approach quality education.

WE HELP YOU IN CHANGING YOUR EDUCATION PROVIDER FOR A BETTER FUTURE!

Contact us today to solve this issue permanently. We’ll help you out with a list of Education Providers studying with whom you won’t have to worry much about your fee.

Register Now - http://wescholarship.com.au/contact-us/

Contact us at: +61 450 732 207, +61 2 8041 0696
Email: info@wescholarship.com.au